refactor(playback): extract Auto browse tree/voice/resolver under ≤1000 LOC
Split AutoBrowseLibraryCallback into AutoBrowseTreeBuilder,
AutoVoiceSearchHandler, and AutoMediaResolver while keeping the
MediaLibraryService callback as a thin façade.
