+++ core/playback/src/main/java/cx/aswin/boxlore/core/data/service/auto/AutoBrowseContract.kt
@@ -151,3 +151,8 @@ internal object AutoBrowseContract {
             putInt(MediaConstants.EXTRAS_KEY_CONTENT_STYLE_PLAYABLE, playable)
         }
 }
+
+internal fun String.stripEpisodePrefix(): String =
+    removePrefix(AutoBrowseContract.LEARN_PREFIX)
+        .removePrefix(AutoBrowseContract.EPISODE_PREFIX)
+        .removePrefix(AutoBrowseContract.QUEUE_PREFIX)
